Relocate processing from MarathonStore to ServiceTree
