+++ src/js/stores/MarathonStore.js
@@ -439,32 +439,6 @@ class MarathonStore extends GetSetBaseStore {
     return this.get('apps')[name];
   }
 
-  getServiceNameFromTaskID(taskID) {
-    let serviceName = taskID.split('.')[0].split('_');
-    return serviceName[serviceName.length - 1];
-  }
-
-  getServiceFromTaskID(taskID) {
-    let service = this.get('apps')[this.getServiceNameFromTaskID(taskID)];
-
-    if (service == null) {
-      return null;
-    }
-
-    return new Service(service.snapshot);
-  }
-
-  getTaskFromTaskID(taskID) {
-    let service = this.getServiceFromTaskID(taskID);
-    if (service == null || service.tasks == null || !service.tasks.length) {
-      return null;
-    }
-
-    return service.tasks.find(function (task) {
-      return task.id === taskID;
-    });
-  }
-
   processMarathonInfoRequest(info) {
     this.set({info});
     this.emit(MARATHON_INSTANCE_INFO_SUCCESS);
